Split the YOLOX Hooks
